Move resource diff and visitor logic to primary library directory
